diracx-client/pyproject.toml
[project]
name = "diracx-client"
description = "TODO"
readme = "README.md"
requires-python = ">=3.11"
keywords = []
license = { text = "GPL-3.0-only" }
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Distributed Computing",
]
dependencies = ["azure-core", "diracx-core", "isodate", "httpx"]
dynamic = ["version"]

[project.optional-dependencies]
testing = ["diracx-testing"]
types = []

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
root = ".."

[tool.hatch.build.targets.wheel]
packages = ["src/diracx"]

[tool.hatch.build.targets.wheel.force-include]
"src/_diracx_client_importer.pth" = "_diracx_client_importer.pth"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "-v",
    "--cov=diracx.client",
    "--cov-report=term-missing",
    "-pdiracx.testing",
    "--import-mode=importlib",
]
asyncio_mode = "auto"
markers = [
    "enabled_dependencies: List of dependencies which should be available to the FastAPI test client",
]
